refactor: StatisticsScreen 내부 컨텐츠 분리
